Split test file. Difficulties while reading final code. Fixing some stuff break other test. The main challenge seems to be typedef.

+++ lib/src/check_white_then_newline.c
@@ -10,11 +10,10 @@
 
 bool			check_white_then_newline(t_parsing	*p,
 						 const char	*code,
-						 int		pos,
+						 ssize_t	pos,
 						 bool		statement)
 {
-  
-  int			i = pos;
+  ssize_t		i = pos;
 
   if (p->single_instruction_per_line.active == false)
     return (true);
